chore : 엔티티 기본 틀 작성 - rating_avg 컬럼의 정밀도를 확보 - order 필드에 nullable=false와 optional=false 제약을 추가 - Order 도메인 배송지 주소 : columnDefinition = Text 제거 후, Varchar(255)로 수정 - enum 분리
