+++ src/main/java/com/nextdaydelivery/ai_response/domain/entity/AiResponse.java
