+++ src/main/java/com/nextdaydelivery/cart/domain/enums/CartStatus.java
@@ -1,0 +1,6 @@
+package com.nextdaydelivery.cart.domain.enums;
+
+// 장바구니 상태를 위한 ENUM
+public enum CartStatus {
+    ACTIVE, INACTIVE, COMPLETED
+}
