+++ src/main/java/com/nextdaydelivery/delivery/domain/enums/DeliveryStatus.java
@@ -1,0 +1,10 @@
+package com.nextdaydelivery.delivery.domain.enums;
+
+/**
+ * 배달 상태 관리를 위한 ENUM
+ */
+public enum DeliveryStatus {
+    DELIVERY_PENDING,   // 배달 대기
+    DELIVERY_ING,       // 배달 중
+    DELIVERY_COMPLETED  // 배달 완료
+}
