+++ src/main/java/com/nextdaydelivery/payment/domain/entity/enums/PaymentMethod.java
@@ -1,0 +1,8 @@
+package com.nextdaydelivery.payment.domain.entity.enums;
+
+/**
+ * 결제 방법을 위한 ENUM
+ */
+public enum PaymentMethod {
+    CARD
+}
