+++ src/main/java/com/nextdaydelivery/payment/domain/entity/enums/PaymentStatus.java
@@ -1,0 +1,10 @@
+package com.nextdaydelivery.payment.domain.entity.enums;
+
+/**
+ * 결제 상태를 위한 ENUM
+ */
+public enum PaymentStatus {
+    CANCEL,
+    COMPLETE,
+    FAILED
+}
