+++ src/main/java/com/nextdaydelivery/review/domain/entity/enums/ReviewStatus.java
@@ -1,0 +1,9 @@
+package com.nextdaydelivery.review.domain.entity.enums;
+
+/**
+ * 리뷰 상태 관리를 위한 ENUM
+ */
+public enum ReviewStatus {
+    VISIBLE, // 표시
+    HIDDEN   // 숨김
+}
